requirements.txt
torch==1.12.0
hydra-core==1.1.2
tensorboard
matplotlib